Added an input system using `gainput`
I initially turned to gainput because I was having a lot of trouble with
SDL_Joystick. Turns out that I had neglected to give my app permissions
to read USB devices in the xcode build settings. So maybe I'll go back
to SDL_Joystick to remove the dependancy (the input is all abstracted
away from the rest of the engine anyway) but it's possible that gainput
is more feature-rich anyway and I'll want to keep it.

Something to consider: how to deal with the start/stop press stuff in my
input system. It's probably the right solution to just have a bunch of
bools for all the info that gainput will give us and just pool
everything every frame, but that seems wasteful. Then again, simple =
good as Jon Blow would say