Renamed the struct "Component" to "ECSComponent" for compatability
This is totally crazy but "Component" is a typedef in XCTest and so I
can't use it in my own code. Maybe there is a way around it but I would
rather not spend my time digging through the XCode settings so I just
made this quick rename fix. *grumble grumble grumble*