First crack at an EntityAdmin test suite
Two bugs I totally would not have caught if not for this:
- destroying entities didn't remove the associated entity-to-components
map
- creating two entities with the same ID would kind of just quietly
fail, now the interface is changed so that you can
`tryCreateEntity(entityID eID)` and if it fails, you will get a nullptr
back. Obviously if you don't check that you don't have a nullptr (I
don't yet in the deserialization code) this doesn't help, but now the
interface is there for it.

You can also now get entity pointers, although why this would ever be
needed by a user isn't all that clear, but it made testing a lot easier.
It's possible this is bad practice but its also possible that it might
be useful to me at some point.