Compile-time optimization pass
Removed unnecessary includes and the like.

I added #pragma once in front of all my header files in addition to the
standard include guards because the internet told me to, but I have no
actual idea if that gave me any speedup benefit.

The one big thing I didn't attempt yet is to create a forward
declaration version of EntityAdmin.hpp to make it less of a choke
point for the compiler.