Added a wholly incomplete editor move mechanic
Right now it defaults to the X axis and doesn't actually check your
mouse was over the move handle and the math isn't the way everyone else
does it but doesn't actually bother me that much for how I would want to
do it, but maybe that is just because of the wacom. In any case, it was
great to see stuff actually move.