Dragging objects around in the world now works!
It isn't the way that every other 3D editor does it, but I might just
like this way better with my wacom. I don't think the math would be too
difficult to get it to be the standard way though. Something to consider
as I play around with it more.

I am really starting to feel the power of what I've built before come in
to accelerate the process of getting new stuff off the ground. The
library of code I've created underneath me is really exciting.