Added orientation imgui widget to Transform
Note that it is kind of a cheat -- Transforming between quats and euler
angles is not nearly as simple as my code makes it out to be (there are
two possible answers every time, and I'm not even sure how glm chooses
the one it wants to use) but it is functional enough for right now.

I looked into https://github.com/BrutPitt/imGuIZMO.quat but decided I
already had enough of other people's code in this project, at least for
now.